+++ navyu_path_planner/include/navyu_path_planner/node.hpp
