+++ navyu_path_planner/launch/navyu_global_planner.launch.py
@@ -25,14 +25,14 @@ def generate_launch_description():
     use_sim_time = LaunchConfiguration("use_sim_time", default="true")
 
     navyu_global_planner_config = PathJoinSubstitution(
-        [FindPackageShare("navyu_planner"), "config", "navyu_global_planner_params.yaml"]
+        [FindPackageShare("navyu_path_planner"), "config", "navyu_global_planner_params.yaml"]
     )
 
     return LaunchDescription(
         [
             DeclareLaunchArgument("use_sim_time", default_value="true"),
             Node(
-                package="navyu_planner",
+                package="navyu_path_planner",
                 executable="navyu_global_planner_node",
                 name="navyu_global_planner_node",
                 output="screen",
